Refactoring: split workflow stages to separate files
